+++ elements/share/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/share
 
+## 3.5.4
+
+### Patch Changes
+
+- [`af6e73a1e17`](https://bitbucket.org/atlassian/atlassian-frontend/commits/af6e73a1e17) - Bumping dependencies via Renovate:
+
+  - react-loosely-lazy
+
 ## 3.5.3
 
 ### Patch Changes